backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
supabase==2.17.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Vector Database
pinecone==4.0.0

# Web Scraping and Data Extraction
crawl4ai==0.3.4
aiohttp==3.9.1
beautifulsoup4==4.12.2
feedparser==6.0.10
requests==2.31.0

# Serper.dev Integration
httpx==0.25.2

# AI and ML
openai==1.3.7
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2

# Data Processing
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2
pytz==2023.3

# Task Queue and Background Jobs
celery==5.3.4
redis==5.0.1

# Monitoring and Logging
loguru==0.7.2
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Environment and Config
python-dotenv==1.0.0
pyyaml==6.0.1

# Data Validation and Processing
jsonschema==4.20.0
python-multipart==0.0.6

# Caching
redis==5.0.1
aioredis==2.0.1

# PDF and Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Image Processing
Pillow==10.1.0

# URL and Domain Processing
tldextract==5.1.1
furl==2.1.3

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6

# Authentication (for future implementation)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6